Split ProductionService/Controller into Machine/Production. Also components controller
